+++ sdk/relay/arm-relay/src/api/hybridConnections/operations.ts
@@ -1,30 +1,30 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
-import { RelayAPIContext as Client } from "../index.js";
-import {
-  errorResponseDeserializer,
+import type { RelayAPIContext as Client } from "../index.js";
+import type {
   AuthorizationRule,
-  authorizationRuleSerializer,
-  authorizationRuleDeserializer,
   _AuthorizationRuleListResult,
-  _authorizationRuleListResultDeserializer,
   AccessKeys,
-  accessKeysDeserializer,
   RegenerateAccessKeyParameters,
-  regenerateAccessKeyParametersSerializer,
   HybridConnection,
-  hybridConnectionSerializer,
-  hybridConnectionDeserializer,
   _HybridConnectionListResult,
-  _hybridConnectionListResultDeserializer,
 } from "../../models/models.js";
 import {
-  PagedAsyncIterableIterator,
-  buildPagedAsyncIterator,
-} from "../../static-helpers/pagingHelpers.js";
+  errorResponseDeserializer,
+  authorizationRuleSerializer,
+  authorizationRuleDeserializer,
+  _authorizationRuleListResultDeserializer,
+  accessKeysDeserializer,
+  regenerateAccessKeyParametersSerializer,
+  hybridConnectionSerializer,
+  hybridConnectionDeserializer,
+  _hybridConnectionListResultDeserializer,
+} from "../../models/models.js";
+import type { PagedAsyncIterableIterator } from "../../static-helpers/pagingHelpers.js";
+import { buildPagedAsyncIterator } from "../../static-helpers/pagingHelpers.js";
 import { expandUrlTemplate } from "../../static-helpers/urlTemplate.js";
-import {
+import type {
   HybridConnectionsListByNamespaceOptionalParams,
   HybridConnectionsDeleteOptionalParams,
   HybridConnectionsCreateOrUpdateOptionalParams,
@@ -36,12 +36,8 @@ import {
   HybridConnectionsCreateOrUpdateAuthorizationRuleOptionalParams,
   HybridConnectionsGetAuthorizationRuleOptionalParams,
 } from "./options.js";
-import {
-  StreamableMethod,
-  PathUncheckedResponse,
-  createRestError,
-  operationOptionsToRequestParameters,
-} from "@azure-rest/core-client";
+import type { StreamableMethod, PathUncheckedResponse } from "@azure-rest/core-client";
+import { createRestError, operationOptionsToRequestParameters } from "@azure-rest/core-client";
 
 export function _listByNamespaceSend(
   context: Client,