+++ sdk/relay/arm-relay/src/classic/index.ts
@@ -6,4 +6,4 @@ export type { NamespacesOperations } from "./namespaces/index.js";
 export type { OperationsOperations } from "./operations/index.js";
 export type { PrivateEndpointConnectionsOperations } from "./privateEndpointConnections/index.js";
 export type { PrivateLinkResourcesOperations } from "./privateLinkResources/index.js";
-export type { WCFRelaysOperations } from "./wcfRelays/index.js";
+export type { wCFRelaysOperations } from "./wCFRelays/index.js";